+++ test/integration/targets/filter_core/tasks/main.yml
@@ -136,65 +136,6 @@
     - "'Ansible - くらとみ\n' | b64encode(encoding='utf-16-le') == 'QQBuAHMAaQBiAGwAZQAgAC0AIABPMIkwaDB/MAoA'"
     - "'QQBuAHMAaQBiAGwAZQAgAC0AIABPMIkwaDB/MAoA' | b64decode(encoding='utf-16-le') == 'Ansible - くらとみ\n'"
 
-- name: Test random_mac filter bad argument type
-  debug:
-    var: "0 | random_mac"
-  register: _bad_random_mac_filter
-  ignore_errors: yes
-
-- name: Verify random_mac filter showed a bad argument type error message
-  assert:
-    that:
-      - _bad_random_mac_filter is failed
-      - "_bad_random_mac_filter.msg is match('Invalid value type (.*int.*) for random_mac .*')"
-
-- name: Test random_mac filter bad argument value
-  debug:
-    var: "'dummy' | random_mac"
-  register: _bad_random_mac_filter
-  ignore_errors: yes
-
-- name: Verify random_mac filter showed a bad argument value error message
-  assert:
-    that:
-      - _bad_random_mac_filter is failed
-      - "_bad_random_mac_filter.msg is match('Invalid value (.*) for random_mac: .* not hexa byte')"
-
-- name: Test random_mac filter prefix too big
-  debug:
-    var: "'00:00:00:00:00:00' | random_mac"
-  register: _bad_random_mac_filter
-  ignore_errors: yes
-
-- name: Verify random_mac filter showed a prefix too big error message
-  assert:
-    that:
-      - _bad_random_mac_filter is failed
-      - "_bad_random_mac_filter.msg is match('Invalid value (.*) for random_mac: 5 colon.* separated items max')"
-
-- name:  Verify random_mac filter
-  assert:
-    that:
-            - "'00' | random_mac is match('^00:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]$')"
-            - "'00:00' | random_mac is match('^00:00:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]$')"
-            - "'00:00:00' | random_mac is match('^00:00:00:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]$')"
-            - "'00:00:00:00' | random_mac is match('^00:00:00:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]:[a-f0-9][a-f0-9]$')"
-            - "'00:00:00:00:00' | random_mac is match('^00:00:00:00:00:[a-f0-9][a-f0-9]$')"
-            - "'00:00:00' | random_mac != '00:00:00' | random_mac"
-
-- name: Verify random_mac filter with seed
-  assert:
-    that:
-      - "'00:00:00' | random_mac(seed='test') == '00:00:00' | random_mac(seed='test')"
-      - "'00:00:00' | random_mac(seed='test') != '00:00:00' | random_mac(seed='another_test')"
-
-- name: Ensure dict2items works with hostvars
-  debug:
-    msg: "{{ item.key }}"
-  loop: "{{ hostvars|dict2items }}"
-  loop_control:
-    label: "{{ item.key }}"
-
 - name: Ensure combining two dictionaries containing undefined variables provides a helpful error
   block:
     - set_fact:
