+++ test/integration/targets/filter_random_mac/aliases
@@ -1,0 +1,3 @@
+shippable/posix/group2
+skip/python2.6  # filters are controller only, and we no longer support Python 2.6 on the controller
+skip/aix
